requirements.txt
asgiref==3.2.5
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
postgis==1.0.4
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
